QnA Assist REST API/requirements.txt
fastapi
uvicorn
openai
python-dotenv
snowflake-snowpark-python
snowflake-connector-python
snowflake-sqlalchemy
cryptography
SQLAlchemy